Config: Modify locale resources relative path to point to root
We are currently copying the locale resources to root. It seems like
Carthage does not support custom subdirectories inside frameworks, and
doing as this project does officially results to a compile error. This,
along with the failing tests are the things blocking us from sending a
pull request to mainline.